commit b325f34d6fd056178a119182f410f157d35a9ea6
    12781 /etc/ppt_aliases needs to be in the boot archive
commit 282a8ecb1f4aca0718d89ef1299b5928e5405bca
    12736 Add DISCARD/TRIM support to bhyve
commit aee33e58f237f0d3cfeab24295f491937f0f7c79
    12766 clean up bhyve lints
commit 00509c829a7bda32eeabfd272e5e6307ed93079e
    12784 Fix errors in plockstat man page
commit c376fe93141769de29a07813de2dfb2c8c4b8481
    12745 man page typos
commit 84659b24a533984de271059abf9a1092835d15a9
    12735 bhyve upstream sync 2019 Sept
commit cf3ec608f736765ec9852eed5e611848a25de9a4
    12768 12392 regressed ftello64 behavior
commit 165c5c6fe7d6c7a95878c8a3aae7da65d1da1d90
    12774 Resilver restarts unnecessarily when it encounters errors
commit 9c3024a3457d2d1269be18124a1ac69e33000da7
    12682 want mdb-bhyve module
commit 88d6421c15f1c4daea9c8128e715662091258054
    12779 mdb str_flags is missing STRXPG4TTY
commit 3579e3c35b601b3f66174315be3026281487541a
    12776 libbe: missing typedef in libbe.h
commit eb9a1df2aeb866bf1de4494433b6d7e5fa07b3ae
    12680 want PCI pass-thru in bhyve

	Conflicts:
	exception_lists/packaging
	usr/src/cmd/bhyve/bhyverun.c
	usr/src/cmd/mdb/i86pc/modules/unix/unix.c
	usr/src/cmd/pptadm/Makefile
	usr/src/lib/libppt/Makefile
	usr/src/lib/libppt/Makefile.com
	usr/src/lib/libvmm/Makefile
	usr/src/lib/libvmm/Makefile.com
	usr/src/uts/i86pc/Makefile.i86pc
	usr/src/uts/i86pc/io/vmm/vm/vm_page.h
	usr/src/uts/i86pc/ppt/Makefile
	usr/src/uts/i86pc/vmm/Makefile
	usr/src/uts/intel/ia32/ml/modstubs.s

+++ usr/src/lib/libc/port/stdio/ftell.c
@@ -91,7 +91,7 @@ ftell_common(FILE *iop)
 		tres += adjust;
 
 	FUNLOCKFILE(lk);
-	return ((long)tres);
+	return (tres);
 }
 
 long
